frontend/requirements.txt
streamlit==1.39.0
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
streamlit-option-menu==0.4.0
streamlit-extras==0.4.7
streamlit-ace==0.1.1